Restructure code to support Home Assistant integration
The `spelldetector` module has been renamed to `magic`. The classes have been split into `SpellDetector`, `SpellClassifier`, & `Spellcaster` to make it clear which parts are doing what.

The `hass` module contains a client that connects to the Home Assistant REST API and the associated triggers that are available based on which spell has been cast.